Adding Identifiable extensions for SwiftUI applications

+++ Sources/KukaiCoreSwift/Models/BakingBad/BetterCallDevTokenMetadata.swift
@@ -13,7 +13,7 @@ public class BetterCallDevTokenMetadata: Codable {
 	
 	public let contract: String
 	public let network: String
-	public let token_id: Int
+	public let token_id: Decimal
 	public let symbol: String?
 	public let name: String?
 	public let decimals: Int?
